refactor: Move epoch aggregation classes to a separate aggregator module "epoch-aggr"
