+++ aggregates/epoch-aggr/src/main/java/com/bloxbean/cardano/yaci/store/epochaggr/storage/impl/EpochStorageImpl.java
@@ -1,0 +1,35 @@
+package com.bloxbean.cardano.yaci.store.epochaggr.storage.impl;
+
+import com.bloxbean.cardano.yaci.store.epochaggr.domain.Epoch;
+import com.bloxbean.cardano.yaci.store.epochaggr.storage.EpochStorage;
+import com.bloxbean.cardano.yaci.store.epochaggr.storage.impl.mapper.EpochMapper;
+import com.bloxbean.cardano.yaci.store.epochaggr.storage.impl.model.EpochEntity;
+import com.bloxbean.cardano.yaci.store.epochaggr.storage.impl.repository.EpochRepository;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Optional;
+
+@RequiredArgsConstructor
+public class EpochStorageImpl implements EpochStorage {
+
+    private final EpochRepository epochRepository;
+    private final EpochMapper epochMapper;
+
+    @Override
+    public void save(Epoch epoch) {
+        EpochEntity updatedEpochEntity = epochRepository.findById(epoch.getNumber())
+                .map(epochEntity -> {
+                    epochMapper.updateEntity(epoch, epochEntity);
+                    return epochEntity;
+                }).orElse(epochMapper.toEpochEntity(epoch));
+
+        epochRepository.save(updatedEpochEntity);
+    }
+
+    @Override
+    public Optional<Epoch> findByNumber(int number) {
+        return epochRepository.findById((long)number)
+                .map(epochEntity -> epochMapper.toEpoch(epochEntity));
+    }
+
+}
