+++ aggregates/epoch-aggr/src/test/java/com/bloxbean/cardano/yaci/store/epochaggr/SpringBootTestApplication.java
@@ -1,0 +1,7 @@
+package com.bloxbean.cardano.yaci.store.epochaggr;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootTestApplication {
+}
