+++ config/application.properties
@@ -113,8 +113,10 @@ logging.file.name= ./logs/yaci-store.log
 # For test networks, it is fine to keep this value low like 300 (5 minutes).
 # Default value is 14400 (4 hours)
 #################################################
-store.blocks.epoch-calculation-interval=14400
-store.blocks.epoch-calculation-enabled=false
+store.epoch-aggr.enabled=false
+store.epoch-aggr.api-enabled=true
+store.epoch-aggr.epoch-calculation-interval=14400
+store.epoch-aggr.epoch-calculation-enabled=false
 
 #######################################################################
 # Enable / Disable local utxo client
@@ -240,3 +242,9 @@ store.auto-index-management=true
 #####################################
 store.db.batch-size=1000
 store.db.parallel-insert=true
+
+#####################################
+# Log level configuration
+#####################################
+logging.level.com.bloxbean.cardano.yaci.core.protocol.keepalive=INFO
+logging.level.com.bloxbean.cardano.yaci.store.core.service=INFO
