+++ settings.gradle
@@ -45,6 +45,7 @@ include 'stores-api:utxo-api'
 include 'stores-api:governance-api'
 
 include 'aggregates:account'
+include 'aggregates:epoch-aggr'
 
 include ':starters:spring-boot-starter'
 include ':starters:utxo-spring-boot-starter'
@@ -61,6 +62,7 @@ include ':starters:submit-spring-boot-starter'
 include ':starters:remote-spring-boot-starter'
 include ':starters:account-spring-boot-starter'
 include ':starters:governance-spring-boot-starter'
+include ':starters:epoch-aggr-spring-boot-starter'
 
 include ':extensions:utxo-rocksdb'
 include ':extensions:account-rocksdb'
