+++ starters/epoch-aggr-spring-boot-starter/build.gradle
@@ -1,0 +1,21 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter'
+    api project(":aggregates:epoch-aggr")
+
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+}
+
+tasks.named('compileJava') {
+    inputs.files(tasks.named('processResources'))
+}
+
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            pom {
+                name = 'Yaci Store Epoch Aggregator Spring Boot Starter'
+                description = 'Yaci Store Epoch Aggregator Spring Boot Starter'
+            }
+        }
+    }
+}
