+++ starters/epoch-aggr-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.bloxbean.cardano.yaci.store.starter.epochaggr.EpochAggrStoreAutoConfiguration
