+++ stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/BlocksStoreConfiguration.java
@@ -1,11 +1,13 @@
 package com.bloxbean.cardano.yaci.store.blocks;
 
-import com.bloxbean.cardano.yaci.store.blocks.storage.*;
-import com.bloxbean.cardano.yaci.store.blocks.storage.impl.*;
+import com.bloxbean.cardano.yaci.store.blocks.storage.BlockStorage;
+import com.bloxbean.cardano.yaci.store.blocks.storage.BlockStorageReader;
+import com.bloxbean.cardano.yaci.store.blocks.storage.RollbackStorage;
+import com.bloxbean.cardano.yaci.store.blocks.storage.impl.BlockStorageImpl;
+import com.bloxbean.cardano.yaci.store.blocks.storage.impl.BlockStorageReaderImpl;
+import com.bloxbean.cardano.yaci.store.blocks.storage.impl.RollbackStorageImpl;
 import com.bloxbean.cardano.yaci.store.blocks.storage.impl.mapper.BlockMapper;
-import com.bloxbean.cardano.yaci.store.blocks.storage.impl.mapper.EpochMapper;
 import com.bloxbean.cardano.yaci.store.blocks.storage.impl.repository.BlockRepository;
-import com.bloxbean.cardano.yaci.store.blocks.storage.impl.repository.EpochRepository;
 import com.bloxbean.cardano.yaci.store.blocks.storage.impl.repository.RollbackRepository;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -39,20 +41,8 @@ public class BlocksStoreConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public EpochStorage epochStorage(EpochRepository epochRepository, EpochMapper epochMapper) {
-        return new EpochStorageImpl(epochRepository, epochMapper);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
     public BlockStorageReader blockStorageReader(BlockRepository blockReadRepository, BlockMapper blockMapper) {
         return new BlockStorageReaderImpl(blockReadRepository, blockMapper);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public EpochStorageReader epochStorageReader(EpochRepository epochReadRepository, EpochMapper epochMapper) {
-        return new EpochStorageReaderImpl(epochReadRepository, epochMapper);
     }
 
     @Bean
