--- stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/domain/EpochsPage.java
@@ -1,24 +1,0 @@
-package com.bloxbean.cardano.yaci.store.blocks.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class EpochsPage {
-    private long total;
-    private int totalPages;
-
-    private List<Epoch> epochs;
-}
