--- stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/storage/EpochStorage.java
@@ -1,10 +1,0 @@
-package com.bloxbean.cardano.yaci.store.blocks.storage;
-
-import com.bloxbean.cardano.yaci.store.blocks.domain.Epoch;
-
-import java.util.Optional;
-
-public interface EpochStorage {
-    void save(Epoch epoch);
-    Optional<Epoch> findByNumber(int number);
-}
