--- stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/storage/impl/repository/EpochRepository.java
@@ -1,12 +1,0 @@
-package com.bloxbean.cardano.yaci.store.blocks.storage.impl.repository;
-
-import com.bloxbean.cardano.yaci.store.blocks.storage.impl.model.EpochEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface EpochRepository extends JpaRepository<EpochEntity, Long> {
-    Optional<EpochEntity> findTopByOrderByNumberDesc();
-}
