+++ stores/blocks/src/main/resources/db/store/h2/V0_100_1__init.sql
@@ -42,20 +42,6 @@ CREATE INDEX idx_block_slot_leader
 CREATE INDEX idx_block_slot
     ON block(slot);
 
-create table epoch
-(
-    number bigint       not null primary key,
-    block_count         int             null,
-    transaction_count   bigint          null,
-    total_output        numeric(38)     null,
-    total_fees          bigint          null,
-    start_time          bigint          null,
-    end_time            bigint          null,
-    max_slot            bigint          null,
-    create_datetime     timestamp,
-    update_datetime     timestamp
-);
-
 drop table if exists rollback;
 create table rollback
 (
